refactor: Section의 일부분이 thread에서도 쓰여 SectionInfo로 분리 후 재사용
